Finish permissions stuff -- make buttons hidden on frontend if user doesnt have permission to use them

+++ sick-fits/backend/src/resolvers/Mutation.js
@@ -2,7 +2,9 @@ const bcrypt = require("bcryptjs")
 const jwt = require("jsonwebtoken")
 const { randomBytes } = require("crypto")
 const { promisify } = require("util")
+
 const { transport, makeANiceEmail } = require("../mail")
+const { hasPermission } = require("../utils")
 
 const Mutations = {
   async createItem(parent, args, ctx, info) {
@@ -49,10 +51,17 @@ const Mutations = {
     const where = { id: args.id }
 
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title}`)
+    const item = await ctx.db.query.item({ where }, `{ id title user { id }}`)
 
     // 2. check if owner or if user has proper permissions
-    // TODO:
+    const ownsItem = item.user.id === ctx.request.userID
+    const hasPermissions = ctx.request.user.permissions.some(permission =>
+      ["ADMIN", "ITEMDELETE"].includes(permission),
+    )
+
+    if (!ownsItem && !hasPermissions) {
+      throw new Error("You don't have permission to delete this item")
+    }
 
     // 3. Delete it!
     return ctx.db.mutation.deleteItem({ where }, info)
@@ -195,6 +204,29 @@ const Mutations = {
 
     // 7. Return the new user
     return updatedUser
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // Check if logged in
+    if (!ctx.request.userID) {
+      throw new Error("You must be logged in to do this")
+    }
+
+    // check if they have the correct update permissions... permission
+    hasPermission(ctx.request.user, ["ADMIN", "PERMISSIONUPDATE"])
+
+    // update permissions in the db
+    return ctx.db.mutation.updateUser(
+      {
+        data: {
+          permissions: {
+            set: args.permissions,
+          },
+        },
+        where: { id: args.userId },
+      },
+      info,
+    )
   },
 }
